Rename package 'cmd' to 'command'
'cmd' is a well-known go package. So, renamed the local package 'cmd' to
'command'.
