+++ command/command_test.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"os"
